Trying out some shaders. The frayma program (entry point) doesn't work correctly(memory errors), and actually infra might need some re-writes as well

+++ program/frayma.cpp
@@ -32,10 +32,9 @@ int main(int argc, char* argv[])
         DEFAULT_PITCH
     );
 
-
     // Get a shader and build the shader object
     Shader shader;
-    status = (int) shader.compile(
+    status = shader.compile(
             frag_shader,
             vert_shader,
             shader_attrib_name
